pyproject.toml
[project]
name = "shynote"
version = "0.3.6"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.6",
    "pydantic==2.10.6",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
    "google-auth>=2.22.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "requests>=2.31.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.2.1",
    "supabase>=2.0.0",
]
